api_duplicate_provider/requirement.txt
alembic==1.8.1
asttokens==2.0.5
backcall==0.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fuzzywuzzy==0.18.0
greenlet==1.1.2
importlib-metadata==4.12.0
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.0
regex==2022.6.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.3.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.1
